docs: transfer the fullBets functions to a separate module
