+++ roulette/models/menuModel.js
@@ -5,59 +5,45 @@ import { rouletteNumber } from './consts/rouletteNumber.js';
 import { rltPos } from './consts/rltPos.js';
 import { rouletteSeries } from './consts/rouletteSeries.js';
 import { oddDozColum } from './consts/oddDozColum.js';
+import { fullBets } from './functions/fullBets.js';
 
 // пересекается ли трек с полем:
 export const trackOverField = true;
 
 
-// шансы, дюжины и колонки:
-// export const oddDozColum = {
-//     dozen_1: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
-//     dozen_2: [13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24],
-//     dozen_3: [25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36],
-//     column_1: [1, 4, 7, 10, 13, 16, 19, 22, 25, 28, 31, 34],
-//     column_2: [2, 5, 8, 11, 14, 17, 20, 23, 26, 29, 32, 35],
-//     column_3: [3, 6, 9, 12, 15, 18, 21, 24, 27, 30, 33, 36],
-//     small: [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18],
-//     big: [19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36],
-//     even: [2, 4, 6, 8, 10, 12, 14, 16, 18, 20, 22, 24, 26, 28, 30, 32, 34, 36],
-//     odd: [1, 3, 5, 7, 9, 11, 13, 15, 17, 19, 21, 23, 25, 27, 29, 31, 33, 35],
-//     red: [1, 3, 5, 7, 9, 12, 14, 16, 18, 19, 21, 23, 25, 27, 30, 32, 34, 36],
-//     black: [2, 4, 6, 8, 10, 11, 13, 15, 17, 20, 22, 24, 26, 28, 29, 31, 33, 35]
-// };
-
-
 // функция загрузки номера:
-export function fullBets(strUp) {
-    // ставка в номер
-    let strup = rouletteNumber[strUp].positions.numb.length * payoutRatios.numb.position * completeBet;
-    console.log("ставка в номер " + strUp + ": " + strup);
-    // ставка на сплиты
-    for (let i = 0; i < rouletteNumber[strUp].positions.split.length; i++) {
-        let position = rouletteNumber[strUp].positions.split[i];
-        let bet = 1 * payoutRatios.split.position * completeBet;
-        console.log("ставка на сплит " + position + ": " + bet);
-    };
-    // ставка на карэ
-    for (let i = 0; i < rouletteNumber[strUp].positions.corner.length; i++) {
-        let position = rouletteNumber[strUp].positions.corner[i];
-        let bet = 1 * payoutRatios.corner.position * completeBet;
-        console.log("ставка на карэ " + position + ": " + bet);
-    };
-    // ставка на стриты
-    for (let i = 0; i < rouletteNumber[strUp].positions.street.length; i++) {
-        let position = rouletteNumber[strUp].positions.street[i];
-        let bet = 1 * payoutRatios.street.position * completeBet;
-        console.log("ставка на стрит " + position + ": " + bet);
-    };
-    // ставка на six_line
-    for (let i = 0; i < rouletteNumber[strUp].positions.six_line.length; i++) {
-        let position = rouletteNumber[strUp].positions.six_line[i];
-        let bet = 1 * payoutRatios.six_line.position * completeBet;
-        console.log("ставка на six_line " + position + ": " + bet);
-    };
-}
-// console.log(fullBets(0));
+// export function fullBets(strUp) {
+//     // ставка в номер
+//     let strup = rouletteNumber[strUp].positions.numb.length * payoutRatios.numb.position * completeBet;
+//     console.log("ставка в номер " + strUp + ": " + strup);
+//     // ставка на сплиты
+//     for (let i = 0; i < rouletteNumber[strUp].positions.split.length; i++) {
+//         let position = rouletteNumber[strUp].positions.split[i];
+//         let bet = 1 * payoutRatios.split.position * completeBet;
+//         console.log("ставка на сплит " + position + ": " + bet);
+//     };
+//     // ставка на карэ
+//     for (let i = 0; i < rouletteNumber[strUp].positions.corner.length; i++) {
+//         let position = rouletteNumber[strUp].positions.corner[i];
+//         let bet = 1 * payoutRatios.corner.position * completeBet;
+//         console.log("ставка на карэ " + position + ": " + bet);
+//     };
+//     // ставка на стриты
+//     for (let i = 0; i < rouletteNumber[strUp].positions.street.length; i++) {
+//         let position = rouletteNumber[strUp].positions.street[i];
+//         let bet = 1 * payoutRatios.street.position * completeBet;
+//         console.log("ставка на стрит " + position + ": " + bet);
+//     };
+//     // ставка на six_line
+//     for (let i = 0; i < rouletteNumber[strUp].positions.six_line.length; i++) {
+//         let position = rouletteNumber[strUp].positions.six_line[i];
+//         let bet = 1 * payoutRatios.six_line.position * completeBet;
+//         console.log("ставка на six_line " + position + ": " + bet);
+//     };
+// }
+
+
+
 // клик по кнопке "Рассчитать":
 calculate.onclick = function () {
     let strUp = document.getElementById('str_up').value;
@@ -136,6 +122,7 @@ export function chipsNeededForABet(arr) {
 }
 // общее число позиций без учёта пересечений:
 // console.log(chipsNeededForABet(selectedNumbers));
+
 // функция подсчёта совпадений (сдача):
 export function countPositions(arr) {
     let total = [];
